Refactor tests related to Github sync
Separated Github syncing related tests into its own test file and made minor adjustments to existing tests. The UserServiceTest is now simplified by removing its dependency on GithubService and transferring the relevant test cases to the newly created GithubSyncServiceTest. Adjustments to UserControllerTest have also been carried out such as changing the @WithMockUser value.
